Cập nhật mới

+++ .vscode/OOP/Tuan7(On thi).cpp/Cau3/hk120192020.cpp
@@ -73,7 +73,7 @@ protected:
     string DiaChiThuaDat;
     double DienTich;
     Ngay NgayCap;
-    double DonGiaThue;
+    long long DonGiaThue;
 
 public:
     DatDai() {}
